+++ SourceFiles/TaskbarWindow.cpp
@@ -1,0 +1,380 @@
+#include "TaskbarWindow.h"
+#include "StartButton.h"
+#include "ClockWidget.h"
+#include "Logger.h"
+#include "resource.h"
+#include <dwmapi.h>
+#include <windowsx.h>
+#include <uxtheme.h>
+#pragma comment(lib, "dwmapi.lib")
+#pragma comment(lib, "shell32.lib")
+#pragma comment(lib, "uxtheme.lib")
+
+HWND g_hTaskbar = NULL;
+HWND g_hNativeTaskbar = NULL;
+HWND g_hTrayNotify = NULL;
+HWND g_hTrayClock = NULL;
+HWND g_hSysPager = NULL;
+HWND g_hReBar = NULL;
+HWND g_hToolbar = NULL;
+
+#define CLASS_NAME L"Shell_TrayWnd"
+#define TRAY_CLASS_NAME L"Elite_SecondaryTrayWnd"
+
+#define IDM_TASKBAR_TOOLBARS        3001
+#define IDM_TASKBAR_CASCADE         3002
+#define IDM_TASKBAR_STACKED         3003
+#define IDM_TASKBAR_SIDEBYSIDE      3004
+#define IDM_TASKBAR_SHOWDESKTOP     3005
+#define IDM_TASKBAR_TASKMGR         3006
+#define IDM_TASKBAR_LOCK            3007
+#define IDM_TASKBAR_PROPERTIES      3008
+#define IDM_START_EXPLORER          3009
+#define IDM_EXIT_ELITETASKBAR       3010
+#define IDM_RESTART_SHELL           3011
+
+UINT g_uTaskbarCreatedMsg = 0;
+OrbState g_orbState = OrbState::Normal;
+bool g_bOrbTrackingMouse = false;
+
+LRESULT CALLBACK TrayNotifyProc(HWND hwnd, UINT uMsg, WPARAM wParam, LPARAM lParam) {
+    return DefWindowProcW(hwnd, uMsg, wParam, lParam);
+}
+
+LRESULT CALLBACK TrayClockProc(HWND hwnd, UINT uMsg, WPARAM wParam, LPARAM lParam) {
+    switch (uMsg) {
+    case WM_CREATE:
+        SetTimer(hwnd, 1, 1000, NULL);
+        return 0;
+    case WM_TIMER:
+        InvalidateRect(hwnd, NULL, FALSE);
+        return 0;
+    case WM_PAINT: {
+        PAINTSTRUCT ps;
+        HDC hdc = BeginPaint(hwnd, &ps);
+        RECT rcClient;
+        GetClientRect(hwnd, &rcClient);
+
+        // Draw Clock text on Glass using BufferedPaint for Alpha preservation
+        BP_PAINTPARAMS params = { sizeof(BP_PAINTPARAMS) };
+        params.dwFlags = BPPF_ERASE;
+        HDC hdcBuffer;
+        HPAINTBUFFER hBufferedPaint = BeginBufferedPaint(hdc, &rcClient, BPBF_TOPDOWNDIB, &params, &hdcBuffer);
+        
+        if (hBufferedPaint) {
+            HTHEME hTheme = OpenThemeData(hwnd, L"Taskbar");
+            if (hTheme) {
+                SYSTEMTIME st;
+                GetLocalTime(&st);
+                wchar_t timeBuf[32];
+                GetTimeFormatW(LOCALE_USER_DEFAULT, TIME_NOSECONDS, &st, NULL, timeBuf, 32);
+                
+                DTTOPTS dttOpts = { sizeof(DTTOPTS) };
+                dttOpts.dwFlags = DTT_COMPOSITED | DTT_TEXTCOLOR | DTT_GLOWSIZE;
+                dttOpts.crText = RGB(255, 255, 255);
+                dttOpts.iGlowSize = 10;
+                
+                // Select a decent font
+                LOGFONTW lf = {0};
+                lf.lfHeight = -14;
+                lf.lfWeight = FW_NORMAL;
+                wcscpy_s(lf.lfFaceName, L"Segoe UI");
+                HFONT hFont = CreateFontIndirectW(&lf);
+                HFONT hOldFont = (HFONT)SelectObject(hdcBuffer, hFont);
+                
+                DrawThemeTextEx(hTheme, hdcBuffer, 0, 0, timeBuf, -1, DT_CENTER | DT_VCENTER | DT_SINGLELINE, &rcClient, &dttOpts);
+                
+                SelectObject(hdcBuffer, hOldFont);
+                DeleteObject(hFont);
+                CloseThemeData(hTheme);
+            }
+            EndBufferedPaint(hBufferedPaint, TRUE);
+        }
+        EndPaint(hwnd, &ps);
+        return 0;
+    }
+    case WM_DESTROY:
+        KillTimer(hwnd, 1);
+        return 0;
+    }
+    return DefWindowProcW(hwnd, uMsg, wParam, lParam);
+}
+
+LRESULT CALLBACK TrayShowDesktopButtonProc(HWND hwnd, UINT uMsg, WPARAM wParam, LPARAM lParam) {
+    switch (uMsg) {
+    case WM_LBUTTONDOWN:
+        // Simulate Win+D
+        keybd_event(VK_LWIN, 0, 0, 0);
+        keybd_event('D', 0, 0, 0);
+        keybd_event('D', 0, KEYEVENTF_KEYUP, 0);
+        keybd_event(VK_LWIN, 0, KEYEVENTF_KEYUP, 0);
+        return 0;
+    case WM_PAINT: {
+        PAINTSTRUCT ps;
+        HDC hdc = BeginPaint(hwnd, &ps);
+        // It's mostly transparent, so we don't draw much, maybe a subtle edge.
+        EndPaint(hwnd, &ps);
+        return 0;
+    }
+    }
+    return DefWindowProcW(hwnd, uMsg, wParam, lParam);
+}
+
+LRESULT CALLBACK WindowProc(HWND hwnd, UINT uMsg, WPARAM wParam, LPARAM lParam) {
+    if (uMsg == g_uTaskbarCreatedMsg && g_uTaskbarCreatedMsg != 0) {
+        // Re-apply AppBar reservation if Explorer restarts
+        APPBARDATA abd = {0};
+        abd.cbSize = sizeof(APPBARDATA);
+        abd.hWnd = g_hTaskbar;
+        SHAppBarMessage(ABM_NEW, &abd);
+        return 0;
+    }
+
+    switch (uMsg) {
+    case WM_CREATE: {
+        DWM_BLURBEHIND bb = {0};
+        bb.dwFlags = DWM_BB_ENABLE;
+        bb.fEnable = TRUE;
+        bb.hRgnBlur = NULL;
+        DwmEnableBlurBehindWindow(hwnd, &bb);
+        return 0;
+    }
+    case WM_PAINT: {
+        PAINTSTRUCT ps;
+        HDC hdc = BeginPaint(hwnd, &ps);
+        RECT rcClient;
+        GetClientRect(hwnd, &rcClient);
+        
+        // Draw Taskbar Background using System Theme (UXTheme)
+        HTHEME hTheme = OpenThemeData(hwnd, L"Taskbar");
+        if (hTheme) {
+            DrawThemeBackground(hTheme, hdc, 1 /*TBP_BACKGROUNDBOTTOM*/, 0, &rcClient, NULL);
+            CloseThemeData(hTheme);
+        } else {
+            // Fallback: Fill with black (Glass handles alpha)
+            FillRect(hdc, &rcClient, (HBRUSH)GetStockObject(BLACK_BRUSH));
+        }
+
+        EndPaint(hwnd, &ps);
+        return 0;
+    }
+    case WM_COMMAND: {
+        int wmId = LOWORD(wParam);
+        switch (wmId) {
+            case IDM_TASKBAR_TASKMGR:
+                ShellExecuteW(NULL, L"open", L"taskmgr.exe", NULL, NULL, SW_SHOWNORMAL);
+                break;
+            case IDM_EXIT_ELITETASKBAR:
+                SendMessageW(hwnd, WM_CLOSE, 0, 0);
+                break;
+            case IDM_START_EXPLORER:
+                ShellExecuteW(NULL, L"open", L"explorer.exe", NULL, NULL, SW_SHOWNORMAL);
+                break;
+            case IDM_RESTART_SHELL:
+                // Restart Explorer Shell
+                ShellExecuteW(NULL, L"open", L"cmd.exe", L"/c taskkill /f /im explorer.exe & start explorer.exe", NULL, SW_HIDE);
+                break;
+            // Properties and others will be implemented in later phases
+        }
+        return 0;
+    }
+    case WM_RBUTTONUP: {
+        int xPos = GET_X_LPARAM(lParam);
+        POINT pt;
+        GetCursorPos(&pt);
+
+        HMENU hMenu = CreatePopupMenu();
+        if (xPos <= 60) {
+            // Start Button Context Menu
+            AppendMenuW(hMenu, MF_STRING, IDM_START_EXPLORER, L"Open Windows Explorer");
+            AppendMenuW(hMenu, MF_STRING, IDM_TASKBAR_PROPERTIES, L"Properties");
+            AppendMenuW(hMenu, MF_SEPARATOR, 0, NULL);
+            AppendMenuW(hMenu, MF_STRING, IDM_EXIT_ELITETASKBAR, L"Exit Elite Taskbar");
+        } else {
+            // Taskbar Context Menu
+            HMENU hToolbars = CreatePopupMenu();
+            AppendMenuW(hMenu, MF_POPUP, (UINT_PTR)hToolbars, L"Toolbars");
+            AppendMenuW(hMenu, MF_SEPARATOR, 0, NULL);
+            AppendMenuW(hMenu, MF_STRING, IDM_TASKBAR_CASCADE, L"Cascade windows");
+            AppendMenuW(hMenu, MF_STRING, IDM_TASKBAR_STACKED, L"Show windows stacked");
+            AppendMenuW(hMenu, MF_STRING, IDM_TASKBAR_SIDEBYSIDE, L"Show windows side by side");
+            AppendMenuW(hMenu, MF_SEPARATOR, 0, NULL);
+            AppendMenuW(hMenu, MF_STRING, IDM_TASKBAR_SHOWDESKTOP, L"Show the desktop");
+            AppendMenuW(hMenu, MF_SEPARATOR, 0, NULL);
+            AppendMenuW(hMenu, MF_STRING, IDM_TASKBAR_TASKMGR, L"Task Manager");
+            AppendMenuW(hMenu, MF_SEPARATOR, 0, NULL);
+            AppendMenuW(hMenu, MF_STRING, IDM_TASKBAR_LOCK, L"Lock the taskbar");
+            AppendMenuW(hMenu, MF_STRING, IDM_TASKBAR_PROPERTIES, L"Properties");
+            AppendMenuW(hMenu, MF_SEPARATOR, 0, NULL);
+            if (GetAsyncKeyState(VK_SHIFT) & 0x8000) {
+                AppendMenuW(hMenu, MF_STRING, IDM_RESTART_SHELL, L"Restart Shell");
+                AppendMenuW(hMenu, MF_SEPARATOR, 0, NULL);
+            }
+            AppendMenuW(hMenu, MF_STRING, IDM_EXIT_ELITETASKBAR, L"Exit Elite Taskbar");
+        }
+
+        SetForegroundWindow(hwnd);
+        TrackPopupMenuEx(hMenu, TPM_LEFTALIGN | TPM_BOTTOMALIGN | TPM_RIGHTBUTTON, pt.x, pt.y, hwnd, NULL);
+        DestroyMenu(hMenu);
+        return 0;
+    }
+    case WM_DESTROY:
+        PostQuitMessage(0);
+        return 0;
+    }
+    return DefWindowProcW(hwnd, uMsg, wParam, lParam);
+}
+
+BOOL CALLBACK MonitorEnumProc(HMONITOR hMonitor, HDC hdcMonitor, LPRECT lprcMonitor, LPARAM dwData) {
+    MONITORINFO mi = {0};
+    mi.cbSize = sizeof(MONITORINFO);
+    if (GetMonitorInfoW(hMonitor, &mi)) {
+        if (mi.dwFlags & MONITORINFOF_PRIMARY) {
+            RECT* primaryRect = (RECT*)dwData;
+            *primaryRect = mi.rcMonitor;
+        }
+    }
+    return TRUE;
+}
+
+bool TaskbarWindow::Initialize(HINSTANCE hInstance) {
+    g_uTaskbarCreatedMsg = RegisterWindowMessageW(L"TaskbarCreated");
+
+    int taskbarHeight = 40; // Default fallback
+
+    // Hide native taskbar if in Replace mode and grab its height to respect user settings (Small/Large)
+    g_hNativeTaskbar = FindWindowW(L"Shell_TrayWnd", NULL);
+    if (g_hNativeTaskbar) {
+        RECT nativeRect = {0};
+        GetWindowRect(g_hNativeTaskbar, &nativeRect);
+        int nativeHeight = nativeRect.bottom - nativeRect.top;
+        if (nativeHeight > 0 && nativeHeight < 100) {
+            taskbarHeight = nativeHeight;
+        }
+        ShowWindow(g_hNativeTaskbar, SW_HIDE);
+    }
+
+    WNDCLASSEXW wc = {0};
+    wc.cbSize = sizeof(WNDCLASSEXW);
+    wc.lpfnWndProc = WindowProc;
+    wc.hInstance = hInstance;
+    wc.lpszClassName = CLASS_NAME;
+    wc.hIcon = LoadIconW(hInstance, MAKEINTRESOURCEW(IDI_MAIN_PROGRAM));
+    wc.hIconSm = LoadIconW(hInstance, MAKEINTRESOURCEW(IDI_MAIN_PROGRAM));
+    wc.hCursor = LoadCursor(NULL, IDC_ARROW);
+    wc.hbrBackground = (HBRUSH)GetStockObject(BLACK_BRUSH);
+    
+    if (!RegisterClassExW(&wc)) {
+        Logger::Log(L"Failed to register Shell_TrayWnd.");
+        return false;
+    }
+
+    // Register child classes
+    WNDCLASSEXW wcChild = {0};
+    wcChild.cbSize = sizeof(WNDCLASSEXW);
+    wcChild.hInstance = hInstance;
+    wcChild.hCursor = LoadCursor(NULL, IDC_ARROW);
+    wcChild.hbrBackground = (HBRUSH)GetStockObject(BLACK_BRUSH); // Black brush is transparent in DWM Glass
+    
+    wcChild.lpfnWndProc = TrayNotifyProc;
+    wcChild.lpszClassName = L"TrayNotifyWnd";
+    RegisterClassExW(&wcChild);
+
+    wcChild.lpfnWndProc = TrayClockProc;
+    wcChild.lpszClassName = L"TrayClockWClass";
+    RegisterClassExW(&wcChild);
+
+    wcChild.lpfnWndProc = TrayShowDesktopButtonProc;
+    wcChild.lpszClassName = L"TrayShowDesktopButtonWClass";
+    RegisterClassExW(&wcChild);
+
+    RECT primaryRect = {0};
+    EnumDisplayMonitors(NULL, NULL, MonitorEnumProc, (LPARAM)&primaryRect);
+    
+    int screenWidth = primaryRect.right - primaryRect.left;
+    int screenHeight = primaryRect.bottom - primaryRect.top;
+    int xPos = primaryRect.left;
+    int yPos = primaryRect.bottom - taskbarHeight;
+
+    g_hTaskbar = CreateWindowExW(
+        WS_EX_TOOLWINDOW | WS_EX_LAYERED,
+        CLASS_NAME,
+        L"", // Empty title for taskbar
+        WS_POPUP | WS_CLIPSIBLINGS | WS_CLIPCHILDREN,
+        xPos, yPos, screenWidth, taskbarHeight,
+        NULL, NULL, hInstance, NULL
+    );
+
+    if (!g_hTaskbar) {
+        Logger::Log(L"Failed to create window HWND.");
+        return false;
+    }
+    
+    // Create children (Start Button removed, rendered directly on parent DC)
+    g_hReBar = CreateWindowExW(0, L"ReBarWindow32", L"", WS_CHILD | WS_VISIBLE | WS_CLIPSIBLINGS | WS_CLIPCHILDREN | RBS_VARHEIGHT | RBS_BANDBORDERS, 
+        45, 0, screenWidth - 260, taskbarHeight, g_hTaskbar, NULL, hInstance, NULL);
+
+    // Tray area
+    g_hTrayNotify = CreateWindowExW(0, L"TrayNotifyWnd", L"", WS_CHILD | WS_VISIBLE, screenWidth - 200, 0, 200, taskbarHeight, g_hTaskbar, NULL, hInstance, NULL);
+    
+    // Hide SysPager and ToolbarWindow32 until Phase 6 when we implement custom rendering
+    g_hSysPager = CreateWindowExW(0, L"SysPager", L"", WS_CHILD, 0, 0, 110, taskbarHeight, g_hTrayNotify, NULL, hInstance, NULL);
+    g_hToolbar = CreateWindowExW(0, L"ToolbarWindow32", L"", WS_CHILD | WS_CLIPSIBLINGS | WS_CLIPCHILDREN, 0, 0, 110, taskbarHeight, g_hSysPager, NULL, hInstance, NULL);
+    
+    // Clock widget width increased to 90px to prevent text cut-off
+    g_hTrayClock = CreateWindowExW(0, L"TrayClockWClass", L"00:00", WS_CHILD | WS_VISIBLE, 110, 0, 90, taskbarHeight, g_hTrayNotify, NULL, hInstance, NULL);
+    
+    // Show Desktop button at far right
+    CreateWindowExW(0, L"TrayShowDesktopButtonWClass", L"", WS_CHILD | WS_VISIBLE, screenWidth - 15, 0, 15, taskbarHeight, g_hTaskbar, NULL, hInstance, NULL);
+
+    // Apply DWM Aero Glass Effect
+    MARGINS margins = { -1, -1, -1, -1 };
+    DwmExtendFrameIntoClientArea(g_hTaskbar, &margins);
+
+    SetLayeredWindowAttributes(g_hTaskbar, 0, 255, LWA_ALPHA);
+    SetWindowPos(g_hTaskbar, HWND_TOPMOST, xPos, yPos, screenWidth, taskbarHeight, SWP_SHOWWINDOW);
+
+    APPBARDATA abd = {0};
+    abd.cbSize = sizeof(APPBARDATA);
+    abd.hWnd = g_hTaskbar;
+    abd.uEdge = ABE_BOTTOM;
+    abd.rc.left = xPos;
+    abd.rc.right = xPos + screenWidth;
+    abd.rc.top = yPos;
+    abd.rc.bottom = yPos + taskbarHeight;
+
+    SHAppBarMessage(ABM_NEW, &abd);
+    SHAppBarMessage(ABM_QUERYPOS, &abd);
+    SHAppBarMessage(ABM_SETPOS, &abd);
+
+    ShowWindow(g_hTaskbar, SW_SHOW);
+    UpdateWindow(g_hTaskbar);
+
+    // Initialize the floating layered Start Button
+    StartButton::Initialize(hInstance, g_hTaskbar);
+    StartButton::SetOrbImageFromResource(hInstance, IDB_START_ORB);
+    StartButton::Show(xPos, yPos, taskbarHeight);
+
+    return true;
+}
+
+void TaskbarWindow::RunMessageLoop() {
+    MSG msg = {0};
+    while (GetMessageW(&msg, NULL, 0, 0)) {
+        TranslateMessage(&msg);
+        DispatchMessageW(&msg);
+    }
+}
+
+void TaskbarWindow::Cleanup() {
+    if (g_hTaskbar) {
+        APPBARDATA abd = {0};
+        abd.cbSize = sizeof(APPBARDATA);
+        abd.hWnd = g_hTaskbar;
+        SHAppBarMessage(ABM_REMOVE, &abd);
+    }
+    
+    if (g_hNativeTaskbar) {
+        ShowWindow(g_hNativeTaskbar, SW_SHOW);
+    }
+}