+++ SourceFiles/resource.h
@@ -1,0 +1,5 @@
+#pragma once
+
+#define IDI_MAIN_PROGRAM 101
+#define IDI_PREFERENCES  102
+#define IDB_START_ORB    103